refactor(core): complete native.rs split — extract focus.rs and tests
Finishes the decomposition started in the previous commit.
native.rs is reduced to 13 lines: SPDX header and the thin
process_response entry point.

Moves:
- call_llm_with_timeout, call_llm_non_streaming, call_llm_with_retry
  → llm_dispatch.rs
- process_response_native_tools, check_doom_loop,
  process_single_native_turn → tier_loop.rs
- handle_tool_result, record_tool_output_outcome,
  process_successful_tool_output, handle_confirmation_required
  → tool_result.rs
- handle_focus_tool, start/complete_focus_tool, rebuild_knowledge_block,
  handle_compress_context, select/apply_messages_for_compression,
  persist_cancelled_tool_results, build_compression_prompt,
  build_gated_defs_for_iteration
  → new focus.rs (417 lines)

Extracts #[cfg(test)] mod tests into
tool_execution/tests/native_tests.rs (existing test directory).

Also adds #![recursion_limit = "256"] to zeph-bench which instantiates
the same deep Agent<BenchmarkChannel> async stack.

Closes #3669
