+++ crates/zeph-bench/src/lib.rs
@@ -1,6 +1,10 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+// Raised from the default 128 to accommodate deeply nested async state machines
+// generated by #[tracing::instrument] on the agent call stack.
+#![recursion_limit = "256"]
+
 //! Benchmark harness for evaluating Zeph agent performance on standardized datasets.
 //!
 //! `zeph-bench` implements the CLI subcommand `zeph bench` and provides the building blocks
