+++ crates/zeph-core/src/agent/tool_execution/llm_dispatch.rs
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use tracing::Instrument;
+#[cfg(test)]
+use zeph_common::text::estimate_tokens;
 use zeph_llm::provider::{
     ChatResponse, LlmProvider, MessagePart, Role, ThinkingBlock, ToolDefinition,
 };
@@ -336,5 +338,200 @@ impl<C: Channel> Agent<C> {
             last.parts = thinking_parts;
             last.rebuild_content();
         }
+    }
+
+    #[cfg(test)]
+    pub(super) async fn call_llm_with_timeout(
+        &mut self,
+    ) -> Result<Option<String>, crate::agent::error::AgentError> {
+        if self.runtime.lifecycle.cancel_token.is_cancelled() {
+            return Ok(None);
+        }
+
+        if let Some(ref tracker) = self.runtime.metrics.cost_tracker
+            && let Err(e) = tracker.check_budget()
+        {
+            self.channel
+                .send(&format!("Budget limit reached: {e}"))
+                .await?;
+            return Ok(None);
+        }
+
+        let query_embedding = match self.check_response_cache().await? {
+            super::CacheCheckResult::Hit(resp) => return Ok(Some(resp)),
+            super::CacheCheckResult::Miss { query_embedding } => query_embedding,
+        };
+
+        let llm_timeout = std::time::Duration::from_secs(self.runtime.config.timeouts.llm_seconds);
+        let start = std::time::Instant::now();
+        let prompt_estimate = self.runtime.providers.cached_prompt_tokens;
+
+        let memcot_state = match self.services.memory.extraction.memcot_accumulator.as_ref() {
+            Some(acc) => acc.current_state().await,
+            None => None,
+        };
+        let dump_id =
+            self.runtime
+                .debug
+                .debug_dumper
+                .as_ref()
+                .map(|d: &crate::debug_dump::DebugDumper| {
+                    let provider_request = if d.is_trace_format() {
+                        serde_json::Value::Null
+                    } else {
+                        self.provider.debug_request_json(
+                            &self.msg.messages,
+                            &[],
+                            self.provider.supports_streaming(),
+                        )
+                    };
+                    d.dump_request(&crate::debug_dump::RequestDebugDump {
+                        model_name: &self.runtime.config.model_name,
+                        messages: &self.msg.messages,
+                        tools: &[],
+                        provider_request,
+                        memcot_state: memcot_state.as_deref(),
+                    })
+                });
+
+        let trace_guard = self.runtime.debug.trace_collector.as_ref().and_then(|tc| {
+            self.runtime
+                .debug
+                .current_iteration_span_id
+                .map(|id| tc.begin_llm_request(id))
+        });
+
+        let llm_span = tracing::info_span!(
+            "llm.turn_call",
+            model = %self.runtime.config.model_name,
+            provider = self.provider.name(),
+        );
+        let result = self
+            .call_llm_non_streaming(
+                llm_timeout,
+                start,
+                prompt_estimate,
+                dump_id,
+                llm_span,
+                query_embedding,
+            )
+            .await;
+
+        if let Some(guard) = trace_guard
+            && let Some(ref mut tc) = self.runtime.debug.trace_collector
+        {
+            let latency = u64::try_from(start.elapsed().as_millis()).unwrap_or(u64::MAX);
+            let (prompt_tokens, completion_tokens) =
+                self.provider.last_usage().unwrap_or((prompt_estimate, 0));
+            tc.end_llm_request(
+                guard,
+                &crate::debug_dump::trace::LlmAttributes {
+                    model: self.runtime.config.model_name.clone(),
+                    prompt_tokens,
+                    completion_tokens,
+                    latency_ms: latency,
+                    streaming: false,
+                    cache_hit: false,
+                },
+            );
+        }
+
+        result
+    }
+
+    #[cfg(test)]
+    async fn call_llm_non_streaming(
+        &mut self,
+        llm_timeout: std::time::Duration,
+        start: std::time::Instant,
+        prompt_estimate: u64,
+        dump_id: Option<u32>,
+        llm_span: tracing::Span,
+        query_embedding: Option<Vec<f32>>,
+    ) -> Result<Option<String>, crate::agent::error::AgentError> {
+        let cancel = self.runtime.lifecycle.cancel_token.clone();
+        let chat_fut = self.provider.chat(&self.msg.messages).instrument(llm_span);
+        let result = tokio::select! {
+            r = tokio::time::timeout(llm_timeout, chat_fut) => r,
+            () = cancel.cancelled() => {
+                tracing::info!("LLM call cancelled by user");
+                self.update_metrics(|m| m.cancellations += 1);
+                self.channel.send("[Cancelled]").await?;
+                return Ok(None);
+            }
+        };
+        match result {
+            Ok(Ok(resp)) => {
+                let elapsed = start.elapsed();
+                let latency = u64::try_from(elapsed.as_millis()).unwrap_or(u64::MAX);
+                let completion_heuristic = estimate_tokens(&resp) as u64;
+                let (final_prompt, final_completion) = self
+                    .provider
+                    .last_usage()
+                    .unwrap_or((prompt_estimate, completion_heuristic));
+                self.update_metrics(|m| {
+                    m.api_calls += 1;
+                    m.last_llm_latency_ms = latency;
+                    m.context_tokens = final_prompt;
+                    m.prompt_tokens += final_prompt;
+                    m.completion_tokens += final_completion;
+                    m.total_tokens = m.prompt_tokens + m.completion_tokens;
+                });
+                self.record_cost_and_cache(final_prompt, final_completion);
+                self.record_successful_task();
+                if let Some(ref recorder) = self.runtime.metrics.histogram_recorder {
+                    recorder.observe_llm_latency(elapsed);
+                }
+                if self.run_response_verification(&resp) {
+                    let _ = self
+                        .channel
+                        .send("[security] Response blocked by injection detection.")
+                        .await;
+                    return Ok(None);
+                }
+                let cleaned = self.scan_output_and_warn(&resp);
+                if let (Some(d), Some(id)) = (self.runtime.debug.debug_dumper.as_ref(), dump_id) {
+                    d.dump_response(id, &cleaned);
+                }
+                let display = self.maybe_redact(&cleaned);
+                self.channel.send(&display).await?;
+                self.store_response_in_cache(&cleaned, query_embedding)
+                    .await;
+                Ok(Some(cleaned))
+            }
+            Ok(Err(e)) => Err(e.into()),
+            Err(_) => {
+                self.channel
+                    .send("LLM request timed out. Please try again.")
+                    .await?;
+                Ok(None)
+            }
+        }
+    }
+
+    #[cfg(test)]
+    pub(in crate::agent) async fn call_llm_with_retry(
+        &mut self,
+        max_attempts: usize,
+    ) -> Result<Option<String>, crate::agent::error::AgentError> {
+        for attempt in 0..max_attempts {
+            match self.call_llm_with_timeout().await {
+                Ok(result) => return Ok(result),
+                Err(e) if e.is_context_length_error() && attempt + 1 < max_attempts => {
+                    tracing::warn!(
+                        attempt,
+                        "LLM context length exceeded, compacting and retrying"
+                    );
+                    let _ = self
+                        .channel
+                        .send_status("context too long, compacting...")
+                        .await;
+                    let _ = self.compact_context().await?;
+                    let _ = self.channel.send_status("").await;
+                }
+                Err(e) => return Err(e),
+            }
+        }
+        unreachable!("loop covers all attempts")
     }
 }
