+++ crates/zeph-core/src/agent/tool_execution/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+mod focus;
 mod layer_hooks;
 mod llm_dispatch;
 mod metrics_compact;
