+++ crates/zeph-core/src/agent/tool_execution/tool_result.rs
@@ -493,4 +493,289 @@ impl<C: Channel> Agent<C> {
             .last_tool_per_skill
             .insert(skill_name, tool_name.to_owned());
     }
+
+    #[cfg(test)]
+    pub(super) async fn handle_tool_result(
+        &mut self,
+        response: &str,
+        result: Result<Option<zeph_tools::executor::ToolOutput>, zeph_tools::executor::ToolError>,
+    ) -> Result<bool, crate::agent::error::AgentError> {
+        use zeph_sanitizer::{ContentSource, ContentSourceKind};
+        use zeph_tools::executor::ToolError;
+        match result {
+            Ok(Some(output)) => self.process_successful_tool_output(output).await,
+            Ok(None) => {
+                self.record_skill_outcomes("success", None, None).await;
+                self.record_anomaly_outcome(AnomalyOutcome::Success).await?;
+                Ok(false)
+            }
+            Err(ToolError::Blocked { command }) => {
+                tracing::warn!("blocked command: {command}");
+                self.channel
+                    .send("This command is blocked by security policy.")
+                    .await?;
+                self.record_anomaly_outcome(AnomalyOutcome::Blocked).await?;
+                Ok(false)
+            }
+            Err(ToolError::ConfirmationRequired { command }) => {
+                self.handle_confirmation_required(response, &command).await
+            }
+            Err(ToolError::Cancelled) => {
+                tracing::info!("tool execution cancelled");
+                self.update_metrics(|m| m.cancellations += 1);
+                self.channel.send("[Cancelled]").await?;
+                Ok(false)
+            }
+            Err(ToolError::SandboxViolation { path }) => {
+                tracing::warn!("sandbox violation: {path}");
+                self.channel
+                    .send("Command targets a path outside the sandbox.")
+                    .await?;
+                self.record_anomaly_outcome(AnomalyOutcome::Error).await?;
+                Ok(false)
+            }
+            Err(e) => {
+                let category = e.category();
+                let err_str = format!("{e:#}");
+                tracing::error!("tool execution error: {err_str}");
+                if let Some(ref d) = self.runtime.debug.debug_dumper {
+                    d.dump_tool_error("legacy", &e);
+                }
+                let kind = FailureKind::from(category);
+                let sanitized_err = self
+                    .services
+                    .security
+                    .sanitizer
+                    .sanitize(&err_str, ContentSource::new(ContentSourceKind::McpResponse))
+                    .body;
+                self.record_skill_outcomes("tool_failure", Some(&err_str), Some(kind.as_str()))
+                    .await;
+                self.record_anomaly_outcome(AnomalyOutcome::Error).await?;
+
+                if !self.services.learning_engine.was_reflection_used()
+                    && self.attempt_self_reflection(&sanitized_err, "").await?
+                {
+                    return Ok(false);
+                }
+
+                self.channel
+                    .send("Tool execution failed. Please try a different approach.")
+                    .await?;
+                Ok(false)
+            }
+        }
+    }
+
+    /// Record skill learning outcomes for a tool output and optionally trigger self-reflection.
+    ///
+    /// Returns `Ok(true)` when the caller should return early (reflection consumed the turn),
+    /// `Ok(false)` to continue, or `Err` on a hard error.
+    #[cfg(test)]
+    async fn record_tool_output_outcome(
+        &mut self,
+        output: &zeph_tools::executor::ToolOutput,
+    ) -> Result<bool, crate::agent::error::AgentError> {
+        if let Some(ref fs) = output.filter_stats {
+            self.record_filter_metrics(fs);
+        }
+        if output.summary.trim().is_empty() {
+            tracing::warn!("tool execution returned empty output");
+            self.record_skill_outcomes("success", None, None).await;
+            return Ok(true);
+        }
+        if output.summary.contains("[error]") || output.summary.contains("[exit code") {
+            let kind = FailureKind::from_error(&output.summary);
+            self.record_skill_outcomes("tool_failure", Some(&output.summary), Some(kind.as_str()))
+                .await;
+            if !self.services.learning_engine.was_reflection_used()
+                && self
+                    .attempt_self_reflection(&output.summary, &output.summary)
+                    .await?
+            {
+                return Ok(true);
+            }
+        } else {
+            self.record_skill_outcomes("success", None, None).await;
+        }
+        Ok(false)
+    }
+
+    #[cfg(test)]
+    async fn process_successful_tool_output(
+        &mut self,
+        output: zeph_tools::executor::ToolOutput,
+    ) -> Result<bool, crate::agent::error::AgentError> {
+        use crate::agent::format_tool_output;
+        use crate::channel::{ToolOutputEvent, ToolStartEvent};
+        use zeph_llm::provider::{Message, MessagePart, Role};
+
+        if self.record_tool_output_outcome(&output).await? {
+            return Ok(false);
+        }
+
+        let tool_call_id = uuid::Uuid::new_v4().to_string();
+        let tool_started_at = std::time::Instant::now();
+        self.channel
+            .send_tool_start(ToolStartEvent {
+                tool_name: output.tool_name.clone(),
+                tool_call_id: tool_call_id.clone(),
+                params: None,
+                parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
+                started_at: std::time::Instant::now(),
+                speculative: false,
+                sandbox_profile: None,
+            })
+            .await?;
+        if let Some(ref d) = self.runtime.debug.debug_dumper {
+            let dump_content = if self.services.security.pii_filter.is_enabled() {
+                self.services
+                    .security
+                    .pii_filter
+                    .scrub(&output.summary)
+                    .into_owned()
+            } else {
+                output.summary.clone()
+            };
+            d.dump_tool_output(output.tool_name.as_str(), &dump_content);
+        }
+        let processed = self.maybe_summarize_tool_output(&output.summary).await;
+        let body = if let Some(ref fs) = output.filter_stats
+            && fs.filtered_chars < fs.raw_chars
+        {
+            format!(
+                "{}\n{processed}",
+                fs.format_inline(output.tool_name.as_str())
+            )
+        } else {
+            processed.clone()
+        };
+        let filter_stats_inline = output.filter_stats.as_ref().and_then(|fs| {
+            (fs.filtered_chars < fs.raw_chars).then(|| fs.format_inline(output.tool_name.as_str()))
+        });
+        let formatted_output = format_tool_output(output.tool_name.as_str(), &body);
+        self.channel
+            .send_tool_output(ToolOutputEvent {
+                tool_name: output.tool_name.clone(),
+                display: self.maybe_redact(&body).to_string(),
+                diff: None,
+                filter_stats: filter_stats_inline,
+                kept_lines: None,
+                locations: output.locations,
+                tool_call_id: tool_call_id.clone(),
+                terminal_id: None,
+                is_error: false,
+                parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
+                raw_response: None,
+                started_at: Some(tool_started_at),
+            })
+            .await?;
+
+        let (llm_body, has_injection_flags) = self
+            .sanitize_tool_output(&processed, output.tool_name.as_str())
+            .await;
+        let user_msg = Message::from_parts(
+            Role::User,
+            vec![MessagePart::ToolOutput {
+                tool_name: output.tool_name.clone(),
+                body: llm_body,
+                compacted_at: None,
+            }],
+        );
+        self.persist_message(
+            Role::User,
+            &formatted_output,
+            &user_msg.parts,
+            has_injection_flags || !self.services.security.flagged_urls.is_empty(),
+        )
+        .await;
+        self.push_message(user_msg);
+        let outcome = if output.summary.contains("[error]") || output.summary.contains("[stderr]") {
+            AnomalyOutcome::Error
+        } else {
+            AnomalyOutcome::Success
+        };
+        self.record_anomaly_outcome(outcome).await?;
+        Ok(true)
+    }
+
+    #[cfg(test)]
+    async fn handle_confirmation_required(
+        &mut self,
+        response: &str,
+        command: &str,
+    ) -> Result<bool, crate::agent::error::AgentError> {
+        use crate::agent::format_tool_output;
+        use crate::channel::{ToolOutputEvent, ToolStartEvent};
+        use zeph_llm::provider::{Message, MessagePart, Role};
+        let prompt = format!("Allow command: {command}?");
+        if self.channel.confirm(&prompt).await? {
+            if let Ok(Some(out)) = self.tool_executor.execute_confirmed_erased(response).await {
+                let confirmed_tool_call_id = uuid::Uuid::new_v4().to_string();
+                let confirmed_started_at = std::time::Instant::now();
+                self.channel
+                    .send_tool_start(ToolStartEvent {
+                        tool_name: out.tool_name.clone(),
+                        tool_call_id: confirmed_tool_call_id.clone(),
+                        params: None,
+                        parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
+                        started_at: std::time::Instant::now(),
+                        speculative: false,
+                        sandbox_profile: None,
+                    })
+                    .await?;
+                if let Some(ref d) = self.runtime.debug.debug_dumper {
+                    let dump_content = if self.services.security.pii_filter.is_enabled() {
+                        self.services
+                            .security
+                            .pii_filter
+                            .scrub(&out.summary)
+                            .into_owned()
+                    } else {
+                        out.summary.clone()
+                    };
+                    d.dump_tool_output(out.tool_name.as_str(), &dump_content);
+                }
+                let processed = self.maybe_summarize_tool_output(&out.summary).await;
+                let formatted = format_tool_output(out.tool_name.as_str(), &processed);
+                self.channel
+                    .send_tool_output(ToolOutputEvent {
+                        tool_name: out.tool_name.clone(),
+                        display: self.maybe_redact(&processed).to_string(),
+                        diff: None,
+                        filter_stats: None,
+                        kept_lines: None,
+                        locations: out.locations,
+                        tool_call_id: confirmed_tool_call_id.clone(),
+                        terminal_id: None,
+                        is_error: false,
+                        parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
+                        raw_response: None,
+                        started_at: Some(confirmed_started_at),
+                    })
+                    .await?;
+                let (llm_body, has_injection_flags) = self
+                    .sanitize_tool_output(&processed, out.tool_name.as_str())
+                    .await;
+                let confirmed_msg = Message::from_parts(
+                    Role::User,
+                    vec![MessagePart::ToolOutput {
+                        tool_name: out.tool_name.clone(),
+                        body: llm_body,
+                        compacted_at: None,
+                    }],
+                );
+                self.persist_message(
+                    Role::User,
+                    &formatted,
+                    &confirmed_msg.parts,
+                    has_injection_flags || !self.services.security.flagged_urls.is_empty(),
+                )
+                .await;
+                self.push_message(confirmed_msg);
+            }
+        } else {
+            self.channel.send("Command cancelled.").await?;
+        }
+        Ok(false)
+    }
 }
